Add the ability to open wikis in the browser (#77)
* WIP: Add the ability to open wikis in the browser

Currently press SELECT and if the app has one it'll *quit UU* and open the browser
I feel like it should be possible to launch the browser without quitting UU as you can launch it from the home menu while the app is running, but that wasn't working.

* Don't quit UU on browser load

NOTE: This now freezes on returning to UU

* Probably fix loading browser

This uses the new aptLaunchSystemApplet from libctru v2.5.0. It's failing to download from pacman for me so I can't test at the moment.

* Fix browser launch

* Add Wiki icon

* Allocate buffer on heap

* paramSize needs to include the null terminator